Implement rubocop-rspec
