+++ .rspec
@@ -1,2 +1,3 @@
 --color
 --profile
+--require spec_helper
