+++ .travis.yml
@@ -5,11 +5,16 @@ rvm:
   - 1.8.7
   - 1.9.3-p551
   - 2.0.0-p648
-  - 2.1.9
   - 2.2.6
   - 2.3.3
   - 2.4.0
+  - 2.5.3
+  - 2.6.0
   - jruby-9.1.12.0
 matrix:
+  include:
+    - rvm: 2.1.10
+      script: bundle exec rake spec
   allow_failures:
     - rvm: 1.8.7
+    - rvm: 1.9.3-p551
