+++ spec/spec_helper.rb
@@ -1,0 +1,1 @@
+require File.expand_path(File.join(File.dirname(__FILE__), '..', 'lib', 'diffy'))
